refactor(epub): split reflow.rs tests into a sibling file
This branch roughly doubled reflow.rs, taking it to about 1500 lines and well
past the size guideline. Extracting the test module the way session.rs already
does — a `#[path]`-included sibling so `super::*` still resolves — brings the
module itself back to ~740 lines with no change to what is tested.

Also reattaches `Laid`'s doc comment, which ended up above `LayoutRequest` when
that type was introduced between the comment and the struct it describes.

The production module is still above the ~500-line target; splitting the
pagination index away from the backend is the obvious next seam, left out of
this branch to keep the diff reviewable.
